% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbknnR-package.R
\docType{package}
\name{bbknnR-package}
\alias{bbknnR}
\alias{bbknnR-package}
\title{bbknnR: Perform Batch Balanced KNN in R}
\description{
A fast and intuitive batch effect removal tool for single-cell data. BBKNN is originally used in the 'scanpy' python package, and now can be used with 'Seurat' seamlessly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ycli1995/bbknnR}
  \item \url{https://github.com/Teichlab/bbknn}
  \item \url{https://bbknn.readthedocs.io/en/latest/}
  \item Report bugs at \url{https://github.com/ycli1995/bbknnR/issues}
}

}
\author{
\strong{Maintainer}: Yuchen Li \email{liyuchen_1995@outlook.com}

}
\keyword{internal}
